requirements.txt
pypdf==3.8.1
azure-ai-formrecognizer==3.2.0
openai==0.27.7
python-dotenv==1.0.0